add permission check to all commands

+++ Dissimilis.WebAPI/Controllers/BoSong/Commands/MultipleBars/DeleteBarsCommand.cs
@@ -1,4 +1,5 @@
-﻿using Dissimilis.WebAPI.Controllers.BoSong.DtoModelsIn;
+﻿using Dissimilis.DbContext.Models.Enums;
+using Dissimilis.WebAPI.Controllers.BoSong.DtoModelsIn;
 using Dissimilis.WebAPI.Controllers.BoSong.DtoModelsOut;
 using Dissimilis.WebAPI.Extensions.Models;
 using Dissimilis.WebAPI.Services;
@@ -28,16 +29,24 @@ namespace Dissimilis.WebAPI.Controllers.BoSong.Commands.MultipleBars
     {
         private readonly SongRepository _songRepository;
         private readonly IAuthService _IAuthService;
+        private readonly _IPermissionCheckerService _IPermissionCheckerService;
 
-        public DeleteBarsCommandHandler(SongRepository songRepository, IAuthService IAuthService)
+
+        public DeleteBarsCommandHandler(SongRepository songRepository, IAuthService IAuthService, _IPermissionCheckerService IPermissionCheckerService)
         {
             _songRepository = songRepository;
             _IAuthService = IAuthService;
+            _IPermissionCheckerService = IPermissionCheckerService;
+
         }
 
         public async Task<UpdatedSongCommandDto> Handle(DeleteBarsCommand request, CancellationToken cancellationToken)
         {
             var song = await _songRepository.GetSongById(request.SongId, cancellationToken);
+
+            var currentUser = _IAuthService.GetVerifiedCurrentUser();
+
+            if (!await _IPermissionCheckerService.CheckPermission(song, currentUser, Operation.Modify, cancellationToken)) throw new UnauthorizedAccessException();
 
             song.DeleteBars(request.Command.FromPosition, request.Command.DeleteLength);
